FEATURE: Update existing redirects with absolute targets
Without this change redirects that have
target uris with absolute paths were not
updated when a new redirect is added
that matches both domain and source path.

Now they are updated when host + sourcepath
match exactly to their target.
Also the target remains absolute and only
the path is updated.